+++ apps/docs/src/app/core/api-files.ts
@@ -68,6 +68,7 @@ export const API_FILES = {
     carousel: ['CarouselComponent', 'CarouselItemComponent', 'CarouselService'],
     combobox: ['ComboboxComponent'],
     checkbox: ['CheckboxComponent'],
+    contentDensity: [],
     datePicker: ['DatePickerComponent'],
     datetimePicker: ['DatetimePickerComponent'],
     dynamicPage: [
@@ -79,6 +80,10 @@ export const API_FILES = {
         'DynamicPageGlobalActionsComponent',
         'DynamicPageFooterComponent',
         'DynamicPageContentComponent'
+    ],
+    facets: [
+      'FacetComponent',
+      'FacetGroupComponent'
     ],
     feedInput: [
         'FeedInputComponent',