+++ apps/docs/src/app/core/core-documentation.routes.ts
@@ -105,6 +105,10 @@ export const ROUTES: Routes = [
                 loadChildren: () => import('./component-docs/dynamic-page/dynamic-page-docs.module').then((m) => m.DynamicPageDocsModule)
             },
             {
+              path: 'facets',
+              loadChildren: () => import('./component-docs/facets/facet-docs.module').then((m) => m.FacetDocsModule)
+            },
+            {
               path: 'feed-list-item',
                 loadChildren: () =>
                     import('./component-docs/feed-list-item/feed-list-item-docs.module').then((m) => m.FeedListItemDocsModule)
@@ -163,6 +167,13 @@ export const ROUTES: Routes = [
                 loadChildren: () =>
                     import('./component-docs/global-config/global-config-docs.module').then(
                         (m) => m.GlobalConfigDocsModule
+                    )
+            },
+            {
+                path: 'content-density',
+                loadChildren: () =>
+                    import('./component-docs/content-density/content-density-docs.module').then(
+                        (m) => m.ContentDensityDocsModule
                     )
             },
             {