+++ apps/docs/src/app/core/documentation/core-documentation.component.ts
@@ -36,6 +36,7 @@ export class CoreDocumentationComponent extends DocumentationBaseComponent {
             { url: 'core/datetime-picker', name: 'Datetime Picker' },
             { url: 'core/dialog', name: 'Dialog' },
             { url: 'core/icon', name: 'Icon' },
+            { url: 'core/facets', name: 'Facets' },
             { url: 'core/feed-list-item', name: 'Feed List Item' },
             { url: 'core/feed-input', name: 'Feed Input' },
             { url: 'core/file-uploader', name: 'File Uploader' },
@@ -103,6 +104,7 @@ export class CoreDocumentationComponent extends DocumentationBaseComponent {
 
         this.utilities = [
             { url: 'core/global-config', name: 'Global Configuration' },
+            { url: 'core/content-density', name: 'Content Density' },
             { url: 'core/infiniteScroll', name: 'Infinite Scroll' },
             { url: 'core/scroll-spy', name: 'Scroll Spy' },
             { url: 'core/theme-switcher', name: 'Theme Switcher API' }