+++ wdio.conf.js
@@ -35,6 +35,7 @@ exports.config = {
         platformA: [
             './e2e/wdio/platform/**/action-bar.e2e-spec.ts',
             './e2e/wdio/platform/**/action-list-item.e2e-spec.ts',
+            './e2e/wdio/platform/**/approval-flow.e2e-spec.ts',
             './e2e/wdio/platform/**/checkbox.e2e-spec.ts',
             './e2e/wdio/platform/**/checkbox-group.e2e-spec.ts',
             './e2e/wdio/platform/**/combobox.e2e-spec.ts',